Title: 좌표사이의 최단경로, Time: 97ms, Memory: 8MB, Status: Passed - CodeTree

+++ 241007/좌표사이의 최단경로/shortest-path-between-coordinates.java
@@ -9,7 +9,7 @@ public class Main {
         int n = Integer.parseInt(split[1]);
         int a = Math.min(h, n);
         int b= Math.max(h, n);
-        int[][] dp = new int[b+1][b+1];
+        long[][] dp = new long[b+1][b+1];
         dp[a][a] = 1;
         for(int i=a; i<=b; i++) {
             for(int j=a; j<=b; j++) {
